+++ client/src/routes/playground/models/+page.svelte
@@ -3,7 +3,6 @@
   import { goto } from '$app/navigation';
   import { env } from '$env/dynamic/public';
   import {
-    Alert,
     Table,
     TableBody,
     TableBodyCell,
@@ -11,24 +10,16 @@
     TableHead,
     TableHeadCell,
     Button,
-    Modal,
-    Label,
-    P,
   } from 'flowbite-svelte';
   import Edit from '../../../components/models/edit.svelte';
+  import AddModel from '../../../components/models/addModel.svelte';
 
   $: models = [];
   let userInfo = {
     authenticated: false,
     token: null
   };
-  let modalOpen = false;
-  let modalModelName = '';
-  let modalModelUri = '';
-  let modalModelPromptTemplate = '';
-  let modelDownload = false;
-  let modelDownloadProgress = 0;
-  let modalErrorMessage = '';
+  let addModel = false;
   let editModel = false;
   let editId = '';
 
@@ -74,29 +65,6 @@
     getModels();
   }
 
-  async function createModel() {
-    if (modalModelName.length === 0) {
-      modalErrorMessage = 'Please enter a name';
-      return;
-    }
-    const req = await fetch(`${env.PUBLIC_API_URL}/models`, {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${userInfo.token}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ name: modalModelName, uri: modalModelUri, promptTemplate: modalModelPromptTemplate })
-    });
-    if (!req.ok) {
-      modalErrorMessage = (await req.json()).message;
-      return;
-    }
-    modalModelName = '';
-    modalModelUri = '';
-    modalModelPromptTemplate = '';
-    modelDownload = true;
-    getModels();
-  }
 </script>
 
 <div class="flex flex-row justify-between p-6">
@@ -104,7 +72,7 @@
     <h1 class="text-xl font-semibold text-gray-900 dark:text-white">Models</h1>
     <p class="text-sm text-gray-500 dark:text-gray-400">Manage your Llama models</p>
   </div>
-  <Button on:click={() => (modalOpen = true)}
+  <Button on:click={() => (addModel = true)}
     ><svg
       xmlns="http://www.w3.org/2000/svg"
       width="20"
@@ -116,7 +84,7 @@
       <path
         d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z"
       />
-    </svg>&nbsp; Install a new model</Button
+    </svg>&nbsp; Add a new model</Button
   >
 </div>
 <div>
@@ -179,53 +147,6 @@
   </Table>
 </div>
 
-<Modal bind:open={modalOpen} size="xs" autoclose={false} class="w-full">
-  <div class="flex flex-col space-y-6">
-    <h3 class="mb-4 text-xl font-medium text-gray-900 dark:text-white">Install a new model</h3>
-    {#if !modelDownload}
-      {#if modalErrorMessage.length > 0}
-        <Alert color="red">
-          <span class="font-medium">An error occured</span>
-          {modalErrorMessage}
-        </Alert>
-      {/if}
-      <Label class="space-y-2">
-        <span>Model name</span>
-        <input
-          name="name"
-          placeholder="Model name"
-          required
-          bind:value={modalModelName}
-          class="block w-full disabled:cursor-not-allowed disabled:opacity-50 focus:border-blue-500 focus:ring-blue-500 dark:focus:border-blue-500 dark:focus:ring-blue-500 bg-gray-50 text-gray-900 dark:bg-gray-600 dark:text-white dark:placeholder-gray-400 border-gray-300 dark:border-gray-500 p-2.5 text-sm rounded-lg"
-          type="text"
-        />
-      </Label>
-      <Label class="space-y-2">
-        <span>Model download URI</span>
-        <input
-          name="repository"
-          placeholder="Model download URI"
-          required
-          bind:value={modalModelUri}
-          class="block w-full disabled:cursor-not-allowed disabled:opacity-50 focus:border-blue-500 focus:ring-blue-500 dark:focus:border-blue-500 dark:focus:ring-blue-500 bg-gray-50 text-gray-900 dark:bg-gray-600 dark:text-white dark:placeholder-gray-400 border-gray-300 dark:border-gray-500 p-2.5 text-sm rounded-lg"
-          type="text"
-        />
-      </Label>
-      <Label class="space-y-2">
-        <span>Model prompt template</span>
-        <textarea draggable="false" rows="2" class="block resize-none w-full disabled:cursor-not-allowed disabled:opacity-50 focus:border-blue-500 focus:ring-blue-500 dark:focus:border-blue-500 dark:focus:ring-blue-500 bg-gray-50 text-gray-900 dark:bg-gray-600 dark:text-white dark:placeholder-gray-400 border-gray-300 dark:border-gray-500 p-2.5 text-sm rounded-lg" placeholder="Model prompt template" bind:value={modalModelPromptTemplate}></textarea>
-      </Label>
-      <button
-        type="submit"
-        class="text-center font-medium focus:ring-4 focus:outline-none inline-flex items-center justify-center px-5 py-2.5 text-sm text-white bg-blue-700 hover:bg-blue-800 focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800 rounded-lg w-full1"
-        on:click|preventDefault={createModel}>Install the new model</button
-      >
-    {:else}
-      <P size="base">Please wait while the model is being downloaded and installed.</P>
-      <P size="base">The model will appear in the list after the download complete.</P>
-    {/if}
-  </div>
-</Modal>
-
+<AddModel bind:modalShow={addModel}></AddModel>
 
 <Edit bind:id={editId} bind:modalShow={editModel}></Edit>
